Move canvas auth selectors into auth
