+++ src/state/selectors/canvases.js
@@ -1,12 +1,9 @@
 import { createSelector } from 'reselect';
-import { Utils } from 'manifesto.js/dist-esmodule/Utils';
 import flatten from 'lodash/flatten';
 import CanvasGroupings from '../../lib/CanvasGroupings';
 import MiradorCanvas from '../../lib/MiradorCanvas';
 import { miradorSlice } from './utils';
 import { getWindow } from './getters';
-import { getConfig } from './config';
-import { getAuth } from './auth';
 import { getSequence } from './sequences';
 import { getWindowViewType } from './windows';
 
@@ -208,58 +205,3 @@ export const selectInfoResponse = createSelector(
     && infoResponses[iiifServiceId];
   },
 );
-
-export const selectCanvasAuthService = createSelector(
-  [
-    selectInfoResponse,
-    getCanvas,
-    getConfig,
-    getAuth,
-  ],
-  (infoResponse, canvas, { auth: { serviceProfiles = [] } }, auth) => {
-    let iiifResource;
-    iiifResource = infoResponse && infoResponse.json && { ...infoResponse.json, options: {} };
-
-    if (!iiifResource && canvas) {
-      const miradorCanvas = new MiradorCanvas(canvas);
-      const [image] = miradorCanvas.iiifImageResources;
-
-      iiifResource = image && image.getServices()[0];
-    }
-
-    if (!iiifResource) return undefined;
-
-    const orderedAuthServiceProfiles = serviceProfiles.map(p => p.profile);
-
-    let lastAttemptedService;
-
-    for (const profile of orderedAuthServiceProfiles) {
-      const services = getServices(iiifResource, profile);
-      for (const service of services) {
-        if (!auth[service.id]) return service;
-
-        lastAttemptedService = service;
-
-        if (auth[service.id].isFetching || auth[service.id].ok) return service;
-      }
-    }
-
-    return lastAttemptedService;
-  },
-);
-
-/** */
-export function selectAuthStatus({ auth }, service) {
-  if (!service) return null;
-  if (!auth[service.id]) return null;
-  if (auth[service.id].isFetching) return 'fetching';
-  if (auth[service.id].ok) return 'ok';
-  return 'failed';
-}
-
-/** Get all the services that match a profile */
-function getServices(resource, profile) {
-  const services = Utils.getServices(resource);
-
-  return services.filter(service => service.getProfile() === profile);
-}
